% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{avonet}
\alias{avonet}
\title{Avonet data in resolver vignette}
\format{
An object of class \code{data.frame} with 10661 rows and 33 columns.
}
\usage{
avonet
}
\description{
Avonet data in resolver vignette
}
\keyword{datasets}
