% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{eltontraits}
\alias{eltontraits}
\title{EltonTraits data in resolver vignette}
\format{
An object of class \code{data.frame} with 9995 rows and 40 columns.
}
\usage{
eltontraits
}
\description{
EltonTraits data in resolver vignette
}
\keyword{datasets}
