% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nest_heights}
\alias{nest_heights}
\title{Sheard et al. nest height data in resolver vignette}
\format{
An object of class \code{data.frame} with 11122 rows and 4 columns.
}
\usage{
nest_heights
}
\description{
Sheard et al. nest height data in resolver vignette
}
\keyword{datasets}
