% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resolver.R
\name{resolver}
\alias{resolver}
\title{Resolve messy joins between multiple databases with key variables}
\usage{
resolver(
  databases,
  joiners,
  helpers = NULL,
  cache = TRUE,
  cache_loc = getwd(),
  joiners2 = NULL
)
}
\arguments{
\item{databases}{A list of databases to join together}

\item{joiners}{A list of one column per database to join on}

\item{helpers}{A list of one column per database to display during interactive resolver}

\item{cache}{True/False to write cache to \code{cache_loc}}

\item{cache_loc}{Location to write cache}

\item{joiners2}{A list of one column per database to attempt second join on}
}
\value{
A list of joined databases and the crosswalk used to resolve messy joins
}
\description{
Resolve messy joins between multiple databases with key variables
}
\examples{
\dontrun{
  resolver(
databases = list(taxonomy, avonet, eltontraits, nest_heights),
joiners = list("common_name", "common_name", "English", "common_name"),
helpers = list("scientific_name","Species2","Scientific","scientific_name"),
joiners2 = list("id", "id", "id", "id")
)
}
}
