% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{taxonomy}
\alias{taxonomy}
\title{eBird taxonomy data in resolver vignette}
\format{
An object of class \code{data.frame} with 117 rows and 11 columns.
}
\usage{
taxonomy
}
\description{
eBird taxonomy data in resolver vignette
}
\keyword{datasets}
